+++ src/components/BlogHeader.tsx
@@ -1,0 +1,15 @@
+
+const BlogHeader = () => {
+  return (
+    <div className="text-center mb-12">
+      <h1 className="text-4xl font-bold text-navy mb-4">Blog</h1>
+      <p className="text-lg text-gray-600 max-w-3xl mx-auto">
+        Thoughts, tutorials, and insights on computer science, software development,
+        and emerging technologies. I write about topics I'm passionate about and
+        learning experiences from my projects.
+      </p>
+    </div>
+  );
+};
+
+export default BlogHeader;
